feat(): Hackathon edit page
