+++ src/App.tsx
@@ -45,10 +45,11 @@ import TaskBulkImport from "./modules/Dashboard/modules/Tasks/TaskBulkImport";
 import ZonalDashboard from "./modules/Dashboard/modules/zonalDashboard/zonaldashboard";
 import DistrictDashboard from "./modules/Dashboard/modules/districtDashboard/districtdashboard";
 import Hackathon from "./modules/Dashboard/modules/Hackathon/User/Hackathon";
-import HackathonCreate from "./modules/Dashboard/modules/Hackathon/Admin/HackathonCreate";
+import HackathonCreate from "./modules/Dashboard/modules/Hackathon/Admin/pages/HackathonCreate";
 import KKEMLanding from "./modules/KKEM/modules/KKEMLanding";
 import KKEMAuth from "./modules/KKEM/modules/KKEMAuth";
 import RankCard from "./modules/Embeds/modules/RankCard/pages/RankCard";
+import HackathonEdit from "./modules/Dashboard/modules/Hackathon/Admin/pages/HackathonEdit";
 
 const router = createBrowserRouter([
     // Add redirect from '/' to '/login'
@@ -190,8 +191,12 @@ const router = createBrowserRouter([
                     { path: "url-shortener", element: <UrlShortener /> },
                     { path: "hackathon", element: <Hackathon /> },
                     {
-                        path: "hackathon-management",
+                        path: "hackathon/create",
                         element: <HackathonCreate />
+                    },
+                    {
+                        path: "hackathon/edit/:id",
+                        element: <HackathonEdit />
                     }
                 ]
             }
