+++ src/modules/Dashboard/modules/Hackathon/User/styles.css
@@ -207,3 +207,6 @@
 	white-space: nowrap;
 }
 
+.all {
+	margin-top: 3rem;
+}
